lights/requirements.txt
python-lirc==1.2.0
